Do not override invoice's get_lines_to_pay.
We were overriding this method so that when invoice lines were reconciled to
create line lines to pay, the latter were used. This was implemented because we
wanted to implement a basic workflow were just after the invoice was created,
new moves were created if the payment type was Direct Debit. As we're not going
to implement it anyway, this code is useless and confuses users.
(grafted from 0edcf7df5619b260f5263a98843a78b6833f9547)