Properly set default bank account value in invoices.
The bank account was not being properly set. Now we do it on post if the bank
account field is empty and it is required by the payment type.

In version 4.0 invoices and credit notes are no longer separated. That means
that processing a sale can cause an invoice to become a credit note if
sale_invoice_grouping is used, and this can cause the domain for the bank
account to fail.

So the best strategy seems to delay setting the default bank account until the
very end if the user did not fill it in manually.